Move ToplevelBlockEditable into its own file and add doc comments
